modularized sidebar component, created Sidebar.tsx file, changed functions to typescript, added export defaults in Logs.tsx/Sidebar.tsx
